+++ app/forms/listing_form.rb
@@ -1,10 +1,10 @@
 class ListingForm < BaseForm
   with_options default: nil do
     integer :id
-    record  :person
+    record  :category
     record  :location
+    record  :person
     record  :service_area
-    array   :tag_list, default: []
     string  :description
     string  :state
     string  :type
@@ -12,7 +12,9 @@ class ListingForm < BaseForm
 
   def execute
     Listing.find_or_new(id).tap do |listing|
-      listing.attributes = given_inputs
+      listing.attributes = given_inputs.
+        reject{ |k, _v| k == :category }.
+        merge(tag_list: category.lineage)
     end
   end
 end