+++ app/forms/submission_form.rb
@@ -4,7 +4,7 @@ class SubmissionForm < BaseForm
   with_options default: nil do
     integer :id
     record  :service_area
-    hash    :listing_attributes,   strip: false
+    hash    :listings_attributes,  strip: false
     hash    :location_attributes,  strip: false
     hash    :person_attributes,    strip: false
     string  :form_name
@@ -14,7 +14,7 @@ class SubmissionForm < BaseForm
   def execute
     build_location
     build_person
-    build_listing
+    build_listings
     build_submission_responses
     build_submission
   end
@@ -29,12 +29,15 @@ class SubmissionForm < BaseForm
       @person ||= PersonForm.build person_attributes.merge location: @location
     end
 
-    def build_listing
-      @listing ||= ListingForm.build listing_attributes.merge(
-        person: @person,
-        location: @location,
-        service_area: service_area
-      )
+    def build_listings
+      @listings = (listings_attributes[:categories] || []).map do |category_id|
+        ListingForm.build listings_attributes.merge(
+          category: category_id,
+          location: @location,
+          person: @person,
+          service_area: service_area,
+        )
+      end
     end
 
     def build_submission
@@ -64,7 +67,7 @@ class SubmissionForm < BaseForm
         .merge(
           body: body_json,
           person: @person,
-          listings: [@listing],
+          listings: @listings,
           submission_responses: @submission_responses,
         )
     end