+++ app/javascript/pages/Ask.vue
@@ -34,7 +34,7 @@
     /><SpacerField />
 
     <CategoryFields
-      :fieldNamePrefix="withListingPrefix('tag_list[]')"
+      :fieldNamePrefix="withListingPrefix('categories[]')"
       :categories="configuration.categories"
       :tags="listing.tag_list"
     >
@@ -98,7 +98,7 @@ export default {
     }
   },
   created: function() {
-    this.withListingPrefix = partial(fieldNameWithPrefix, 'submission[listing_attributes]')
+    this.withListingPrefix = partial(fieldNameWithPrefix, 'submission[listings_attributes]')
     this.withPersonPrefix  = partial(fieldNameWithPrefix, 'submission[person_attributes]')
   },
 }