+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Enhancements
+* Ask/Offer form now generates a listing for each category selected #536, #620
+
 ### Bugfixes
 * Ask/Offer form breaks when no custom questions are configured #621, #622